Move archiver.util.is10x to utils.protocols.is10x
